requirements.txt
altgraph==0.17.4
packaging==25.0
pillow==11.3.0
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
setuptools==80.9.0
tk==0.1.0
